requirements.txt
pytest==8.2.2
pytest-cov==5.0.0
bandit==1.7.9
pyyaml==6.0.1
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
strawberry-graphql==0.221.0
httpx==0.27.0
